Refactored class Metrics to use user passed parameters.
Compile time sized stack allocated arrays became
runtime sized heap allocated arrays.

Kept changes to a minimum and syntax close to previous
C style syntax.

Pointers became unique pointers.

get_srcline() was returning a 2d stack array before the refactoring,
and callers used [] syntax to access elements.

Now that the whole 3d srcline array is in the heap contiguously, it's not
possible to return a static 2d array anymore.

A proxy class SrcLine2D was created, and an object of this class is returned by get_srcline() instead.

The [] operator on this object could be overloaded.

This would avoid needing to refactor calling code, but would risk bugs because
callers would have an illusion of working with a C 2d array.

To make the change explicit, I decided to implement a explict get() method for
SrcLine2D. This way callers know they are dealing with a object and won't try
to treat it like a 2d C array.

The cost was to refactor some lines of caller code from [i] to get(i).